GODRIVER-2570 decouple log test helpers
